Add files via upload
added in --stream-response.   This will have ollama stream the response and buffer it.   It will start speaking the response after each sentence.   Its parsing is rudimentary right now, looking for periods.   Future versions may have more sentence logic added so that it can detect list responses and speak the entire line.   I also want the debugging output to print the entire response when it is done.