Add files via upload
A number of improvements have been made.  

--verbose has been added to get better debugging output
--language-model has been added so that you can specify which model you prefer to use.   This defaults to openhermes:7b-mistral-v2.5-q4_K_M but you can use which ever one you want that you have downloaded.  

You can interrupt the audio response by pressing and holding spacebar.   It will detect this and no send any more content to the MacOS say command.   It will go back to listening for keyword.

Also quite a bit more error correcting and some changes to the default greetings.    I have found in general the experience is better with a low pause threshold and a simple keyword like 'bingo'